Added more comments after review. Extracted Channel Config into a separate file.
